Fluff C3 seam 3: extract find + review-note decorations
findDecorations.ts carries the find-in-document plugin (state key +
match/active highlighting); reviewNotes.ts carries the comment-anchor
decorations and the DocumentNoteHighlight type, re-exported from
DocumentRichEditor for DocumentDetail. Both verbatim moves.
